Move most of the streaming concerns into EndpointHtmlRenderer so it can vary its output based on this
